nodejs14 and startInstancesWithRetry into .js file
