+++ serverless.yml
@@ -1,8 +1,8 @@
 service: hammertime
 provider:
   name: aws
-  runtime: nodejs16.x
-  memorySize: 512
+  runtime: nodejs14.x
+  memorySize: 128
   stage: '${env:SLICE}'
   region: '${env:AWS_REGION}'
   deploymentBucket: '${env:DEPLOY_BUCKET}'
